+++ init.lua
@@ -7,8 +7,9 @@
 local STI = {
 	_LICENSE     = "MIT/X11",
 	_URL         = "https://github.com/karai17/Simple-Tiled-Implementation",
-	_VERSION     = "0.14.1.4",
-	_DESCRIPTION = "Simple Tiled Implementation is a Tiled Map Editor library designed for the *awesome* LÖVE framework."
+	_VERSION     = "0.14.1.10",
+	_DESCRIPTION = "Simple Tiled Implementation is a Tiled Map Editor library designed for the *awesome* LÖVE framework.",
+	cache        = {}
 }
 
 local path = (...):gsub('%.init$', '') .. "."
@@ -17,8 +18,10 @@ local Map  = require(path .. "map")
 --- Instance a new map.
 -- @param path Path to the map file.
 -- @param plugins A list of plugins to load.
+-- @param ox Offset of map on the X axis (in pixels)
+-- @param oy Offset of map on the Y axis (in pixels)
 -- @return table The loaded Map.
-function STI.new(map, plugins)
+function STI.new(map, plugins, ox, oy)
 	-- Check for valid map type
 	local ext = map:sub(-4, -1)
 	assert(ext == ".lua", string.format(
@@ -37,9 +40,14 @@ function STI.new(map, plugins)
 	setfenv(map, {})
 	map = setmetatable(map(), {__index = Map})
 
-	map:init(path, plugins)
+	map:init(STI, path, plugins, ox, oy)
 
 	return map
+end
+
+--- Flush image cache.
+function STI:flush()
+	self.cache = {}
 end
 
 return STI