+++ plugins/bump.lua
@@ -79,8 +79,9 @@ return {
 	end,
 
 	--- Draw bump collisions world.
+	-- @params world bump world holding the tiles geometry
 	-- @return nil
-	bump_draw = function(map)
+	bump_draw = function(map, world)
 		for k,collidable in pairs(map.bump_collidables) do
 			love.graphics.rectangle("line",world:getRect(collidable))
 		end